Setup menu bar and menu items html with events
The width of the dropdown menu can't be defined depending on the menu items width because of the position: absolute.
A work around should be found with a position: relative, without enlarging the parent menu in the menu bar.
Another solution would be to show a new undecorated window, but it would be a lot of work and would lead to a lot of issues depending on the os.